requirements.txt
bumpversion==0.5.3
coverage==4.4.2
flake8==3.5.0
flake8-per-file-ignores==0.4
mock==2.0.0
packaging==16.8
pytest-runner>=3.0.0,<3.1.0
pytest>=3.2.0,<3.3.0
readme-renderer>=24.0
sceptre-aws-stackoutput-external-resolver>=1.0
setuptools>=40.6.2
git+git://github.com/sceptre/sceptre-core.git@master#egg=sceptre-core
tox>=2.9.1,<3.0.0
twine>=1.12.1
wheel==0.32.3
